Splits bookmarks_controller by request type. Main file requires split files.
