+++ controllers/bookmarks_controller_post.js
@@ -1,0 +1,8 @@
+const db = require('../models');
+const Sequelize = require('sequelize');
+
+module.exports = app => {
+  app.post("/api/list/add", (req, res) => {
+    console.log(req.body);
+  });
+};
